move all todo logic to index (page.tsx)
