--- app/account/page.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const AccountPage = () => {
-  return <div>AccountPage</div>;
-};
-
-export default AccountPage;
